+++ rigs/bendy_limbs/paw.py
@@ -20,12 +20,12 @@
 
 import bpy
 
-from rigify.rigs.limbs.leg import Rig as LegRig
-from .limb_rigs import BaseLimbBendyRig
+from rigify.rigs.limbs.paw import Rig as PawRig
+from .bendy_limb_rigs import BaseLimbBendyRig
 
 
-class Rig(BaseLimbBendyRig, LegRig):
-    """Bendy human leg rig."""
+class Rig(BaseLimbBendyRig, PawRig):
+    """Bendy Paw rig."""
     pass
 
 
@@ -37,75 +37,51 @@ def create_sample(obj):
     bones = {}
 
     bone = arm.edit_bones.new('thigh.L')
-    bone.head[:] = 0.0980, 0.0124, 1.0720
-    bone.tail[:] = 0.0980, -0.0286, 0.5372
+    bone.head[:] = 0.0000, 0.0017, 0.7379
+    bone.tail[:] = 0.0000, -0.0690, 0.4731
     bone.roll = 0.0000
     bone.use_connect = False
     bones['thigh.L'] = bone.name
     bone = arm.edit_bones.new('shin.L')
-    bone.head[:] = 0.0980, -0.0286, 0.5372
-    bone.tail[:] = 0.0980, 0.0162, 0.0852
+    bone.head[:] = 0.0000, -0.0690, 0.4731
+    bone.tail[:] = 0.0000, 0.1364, 0.2473
     bone.roll = 0.0000
     bone.use_connect = True
     bone.parent = arm.edit_bones[bones['thigh.L']]
     bones['shin.L'] = bone.name
     bone = arm.edit_bones.new('foot.L')
-    bone.head[:] = 0.0980, 0.0162, 0.0852
-    bone.tail[:] = 0.0980, -0.0934, 0.0167
-    bone.roll = 0.0000
+    bone.head[:] = 0.0000, 0.1364, 0.2473
+    bone.tail[:] = 0.0000, 0.0736, 0.0411
+    bone.roll = -0.0002
     bone.use_connect = True
     bone.parent = arm.edit_bones[bones['shin.L']]
     bones['foot.L'] = bone.name
     bone = arm.edit_bones.new('toe.L')
-    bone.head[:] = 0.0980, -0.0934, 0.0167
-    bone.tail[:] = 0.0980, -0.1606, 0.0167
-    bone.roll = -0.0000
+    bone.head[:] = 0.0000, 0.0736, 0.0411
+    bone.tail[:] = 0.0000, -0.0594, 0.0000
+    bone.roll = -3.1416
     bone.use_connect = True
     bone.parent = arm.edit_bones[bones['foot.L']]
     bones['toe.L'] = bone.name
-    bone = arm.edit_bones.new('heel.02.L')
-    bone.head[:] = 0.0600, 0.0459, 0.0000
-    bone.tail[:] = 0.1400, 0.0459, 0.0000
-    bone.roll = 0.0000
-    bone.use_connect = False
-    bone.parent = arm.edit_bones[bones['foot.L']]
-    bones['heel.02.L'] = bone.name
-
 
     bpy.ops.object.mode_set(mode='OBJECT')
     pbone = obj.pose.bones[bones['thigh.L']]
-    pbone.rigify_type = 'limbs_bendy.leg'
+    pbone.rigify_type = 'bendy_limbs.paw'
     pbone.lock_location = (False, False, False)
     pbone.lock_rotation = (False, False, False)
     pbone.lock_rotation_w = False
     pbone.lock_scale = (False, False, False)
     pbone.rotation_mode = 'QUATERNION'
     try:
-        pbone.rigify_parameters.separate_ik_layers = True
+        pbone.rigify_parameters.fk_layers = [False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, True, False, False, False, False, False, False, False, False, False, False, False, False, False, False]
     except AttributeError:
         pass
     try:
-        pbone.rigify_parameters.ik_layers = [False, False, False, False, False, False, False, False, False, False, False, False, False, False, True, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False]
+        pbone.rigify_parameters.tweak_layers = [False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, True, False, False, False, False, False, False, False, False, False, False, False, False, False]
     except AttributeError:
         pass
     try:
-        pbone.rigify_parameters.separate_hose_layers = True
-    except AttributeError:
-        pass
-    try:
-        pbone.rigify_parameters.hose_layers = [False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, True, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False]
-    except AttributeError:
-        pass
-    try:
-        pbone.rigify_parameters.limb_type = "leg"
-    except AttributeError:
-        pass
-    try:
-        pbone.rigify_parameters.fk_layers = [False, False, False, False, False, False, False, False, False, False, False, False, False, False, True, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False]
-    except AttributeError:
-        pass
-    try:
-        pbone.rigify_parameters.tweak_layers = [False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, True, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False]
+        pbone.rigify_parameters.limb_type = "paw"
     except AttributeError:
         pass
     pbone = obj.pose.bones[bones['shin.L']]
@@ -129,13 +105,10 @@ def create_sample(obj):
     pbone.lock_rotation_w = False
     pbone.lock_scale = (False, False, False)
     pbone.rotation_mode = 'QUATERNION'
-    pbone = obj.pose.bones[bones['heel.02.L']]
-    pbone.rigify_type = ''
-    pbone.lock_location = (False, False, False)
-    pbone.lock_rotation = (False, False, False)
-    pbone.lock_rotation_w = False
-    pbone.lock_scale = (False, False, False)
-    pbone.rotation_mode = 'QUATERNION'
+    try:
+        pbone.rigify_parameters.limb_type = "paw"
+    except AttributeError:
+        pass
 
     bpy.ops.object.mode_set(mode='EDIT')
     for bone in arm.edit_bones:
@@ -148,10 +121,3 @@ def create_sample(obj):
         bone.select_head = True
         bone.select_tail = True
         arm.edit_bones.active = bone
-
-    for eb in arm.edit_bones:
-        eb.layers = (False, False, False, False, False, False, False, False, False, False, False, False, False, True, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False)
-
-    arm.layers = (False, False, False, False, False, False, False, False, False, False, False, False, False, True, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False)
-
-    return bones
