+++ rigs/bendy_limbs/paw_rear.py
@@ -132,7 +132,7 @@ def create_sample(obj):
 
     bpy.ops.object.mode_set(mode='OBJECT')
     pbone = obj.pose.bones[bones['thigh.L']]
-    pbone.rigify_type = 'limbs_bendy.paw'
+    pbone.rigify_type = 'bendy_limbs.paw'
     pbone.lock_location = (False, False, False)
     pbone.lock_rotation = (False, False, False)
     pbone.lock_rotation_w = False
